% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_create_api_url}
\alias{pxw_create_api_url}
\title{Convert a link to a database table's home page to an API URL}
\usage{
pxw_create_api_url(url_text)
}
\arguments{
\item{url_text}{A URL to a table as a character string.}
}
\value{
A URL to an API.
}
\description{
Convert a link to a database table's home page to an API URL
}
\examples{
url_text <- paste0(
  "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/",
  "START__AM__AM0210__AM0210A/ArbStatusM/"
)
pxw_create_api_url(url_text)
}
