% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_fetch}
\alias{pxw_fetch}
\title{Retrieve data from a pxweb database}
\usage{
pxw_fetch(
  url_text = url_text,
  filters_list = list(),
  kod_kolumn = NULL,
  lopnr_istallet_for_koder = FALSE
)
}
\arguments{
\item{url_text}{Url to a pxweb table. If the table belongs to Statistics
Sweden, the Institute of Economic Research. Jordbruksverket or Tillväxtanalys
the url will be coverted to an api url.}

\item{filters_list}{A list of which selections should be made from various
parameters. If no selection is made, all variables are retrieved. If a variable
is to be removed, it is specified with variable name = "e", for example Region = "e".}

\item{kod_kolumn}{A vector with names of the columns you also want codes for.}

\item{lopnr_istallet_for_koder}{Some pxweb databases are misconfigured so that
they don't have codes as identifiers, but insted uses serial numbers.
Set to TRUE if these are not identified correctly}
}
\value{
A tibble with the contents of the pxweb table
}
\description{
Retrieve data from a pxweb database
}
\examples{
url_text <- pxw_create_api_url(paste0(
  "https://www.statistikdatabasen.scb.se/",
  "pxweb/sv/ssd/START__AA__AA0003__AA0003B/MotFlyktLanKon/"
))
pxw_fetch(
  url_text = url_text, filters_list = list(Region = "12", Tid = "2022"),
  kod_kolumn = "region"
)
}
