% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_last_period}
\alias{pxw_get_last_period}
\title{Hämta senaste tidpunkt i tabellen (år, månad etc)}
\usage{
pxw_get_last_period(url)
}
\arguments{
\item{url}{URL till tabellen webbsida eller API.}
}
\value{
En textsträng med senaste tidpunkt som finns i tabellen
}
\description{
Hämta senaste tidpunkt i tabellen (år, månad etc)
}
\examples{
url <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_get_last_period(url)
}
