% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_period_intervall}
\alias{pxw_get_period_intervall}
\title{Create a period range for the times in the table}
\usage{
pxw_get_period_intervall(url_text, from_per = NULL, to_per = NULL)
}
\arguments{
\item{url_text}{URL to the table's website or API.}

\item{from_per}{The earliest time in the range.}

\item{to_per}{The latest time in the range.}
}
\description{
Create a period range for the times in the table
}
\examples{
url_text <- paste0(
  "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/",
  "START__AM__AM0210__AM0210A/ArbStatusM/")
pxw_get_period_intervall(url_text, from_per = "2024M01", to_per = "2024M03")
}
