% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_periods}
\alias{pxw_get_periods}
\title{Hämta alla tillgängliga perioder}
\usage{
pxw_get_periods(url)
}
\arguments{
\item{url}{URL till tabellen webbsida eller API.}
}
\value{
En vektor med alla tider som finns i databasen, till exempel alla år det finns data för
}
\description{
Hämta alla tillgängliga perioder
}
\examples{
url <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_get_periods(url, from_per = NULL, to_per = NULL)
}
