% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_region_codes}
\alias{pxw_get_region_codes}
\title{Retrieve geographical variable codes from a pxweb table}
\usage{
pxw_get_region_codes(url_text, niva_nchar = NULL, lan_kod = NULL)
}
\arguments{
\item{url_text}{URL to the table's website or API.}

\item{niva_nchar}{If the table contains mixed geographic levels, specify the
length of the codes to be retrieved.
For example, setting niva_nchar = 2 retrieves only county codes.}

\item{lan_kod}{A vector of county codes. For example, lan_kod = c("10", "12")
retrieves codes only for Blekinge and Skåne.}
}
\value{
A vector with municipality codes or similar.
}
\description{
Retrieve geographical variable codes from a pxweb table
}
\examples{
url_text <- paste0(
  "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/",
  "START__AM__AM0210__AM0210A/ArbStatusM/"
)
pxw_get_region_codes(url_text, niva_nchar = 4, lan_kod = "12")
}
