% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_param_till_lopnr}
\alias{pxw_param_till_lopnr}
\title{Gör om koder till löpnummer i de fall databasen är felkonfigurerad och
använder löpnummer istället för koder, till exempel istället för kommunkoder.}
\usage{
pxw_param_till_lopnr(dfvariabellista, params_lista)
}
\arguments{
\item{dfvariabellista}{En lista på variabler som hämtats med pwx_variabellista(url)}

\item{params_lista}{En lista på de parametrar som ingår i tabellen.}
}
\value{
En tibble över alla parametrar och vilka variabler dessa använder
}
\description{
Gör om koder till löpnummer i de fall databasen är felkonfigurerad och
använder löpnummer istället för koder, till exempel istället för kommunkoder.
}
