% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_skapa_api_url}
\alias{pxw_skapa_api_url}
\title{Gör om en länk till databastabellens hemsida till en api-url}
\usage{
pxw_skapa_api_url(url)
}
\arguments{
\item{url}{}
}
\value{
En API-url
}
\description{
Gör om en länk till databastabellens hemsida till en api-url
}
\examples{
#' url <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_skapa_api_url(url)
}
