% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_variablellista}
\alias{pxw_variablellista}
\title{Skapa en lista över tabellens alla parametrar och deras variabler}
\usage{
pxw_variablellista(url)
}
\arguments{
\item{url}{URL till tabellens webbsida eller API.}
}
\value{
En tibble med alla parametrar och variabler som kan användas.
}
\description{
Skapa en lista över tabellens alla parametrar och deras variabler
}
\examples{
url <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_variablellista(url)
}
