% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_variables_list_mod.R
\name{pxw_variables_list_mod}
\alias{pxw_variables_list_mod}
\title{Adjust variable lists from misconfigured pxweb servers so that they are
following SCB standard}
\usage{
pxw_variables_list_mod(url_str, lopnr_in_values = FALSE)
}
\arguments{
\item{url_str}{string, URL to an pxweb API}

\item{lopnr_in_values}{logical, TRUE if the server don't have categories in
the value param.}
}
\value{
A dataframe with all parameters and variables in use
}
\description{
Adjust variable lists from misconfigured pxweb servers so that they are
following SCB standard
}
\examples{
link <- paste0("https://statistik.tillvaxtanalys.se/PxWeb/pxweb/sv/",
"Tillv\%C3\%A4xtanalys\%20statistikdatabas/Tillv\%C3\%A4xtanalys",
"\%20statistikdatabas__Konkurser\%20och\%20offentliga\%20ackord/",
"konk_ar_lan_bransch_2009.px/")
pxw_variables_list_mod()
}
