% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_visa_parameterlista}
\alias{pxw_visa_parameterlista}
\title{Visa tabellens alla parametrare och queryns struktur}
\usage{
pxw_visa_parameterlista(url)
}
\arguments{
\item{url}{}
}
\value{
En tibble med alla parametrar
}
\description{
Visa tabellens alla parametrare och queryns struktur
}
\examples{
url <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_visa_parameterlista(url)
}
