% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_fetch}
\alias{pxw_fetch}
\title{Hämta data från en pxweb-databas}
\usage{
pxw_fetch(
  url,
  params_lista = list(),
  kod_kolumn = NULL,
  lopnr_istallet_for_koder = FALSE
)
}
\arguments{
\item{url}{Url till en pxwebtabell. Om tabellen tillhör SCB, Konjunkturinstitutet.
Jordbruksverket eller Tillväxtanalys omvandlas url:en till tabellens webbsida
till en api-url.}

\item{params_lista}{en lista med vilka urval som ska göras från olika parametrar.
Om inget urval görs hämtas alla variabler. Om en variabel ska tas bort anges det
med variabelnamn = "e", till exempel Kön = "e".}

\item{kod_kolumn}{En vektor med namn på de kolumner man även vill ha koder för.}

\item{lopnr_istallet_for_koder}{Vissa pxwebdatabaser är felgjorda så att man
har koder som identifierare, utan löpnummer. Ange till TRUE om inte dessa identifieras
korrekt}
}
\value{
En tibble med tabellinnehållet i vald pxweb-tabell
}
\description{
Hämta data från en pxweb-databas
}
\examples{
url = pxw_skapa_api_url("https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AA__AA0003__AA0003B/MotFlyktLanKon/")
dfscb_urval <- pxw_fetch(url, params_lista = list(Region = "12", Tid = "2022"),
    kod_kolumn = "region")
}
