% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_fetch}
\alias{pxw_fetch}
\title{Retrieve data from a pxweb database}
\usage{
pxw_fetch(
  url_text = url_text,
  filters_list = list(),
  kod_kolumn = NULL,
  lopnr_istallet_for_koder = FALSE
)
}
\arguments{
\item{url_text}{URL to a pxweb table. If the table belongs to Statistics
Sweden, the Institute of Economic Research, the Swedish Board of Agriculture,
or Growth Analysis, the URL will be converted to an API URL.}

\item{filters_list}{A list specifying the selections to be made from various
parameters. If no selection is made, all variables are retrieved. To exclude
a variable, specify it with variable name = "e", for example, Region = "e".}

\item{kod_kolumn}{A vector with the names of the columns for which you also
want codes.}

\item{lopnr_istallet_for_koder}{Some pxweb databases are misconfigured and do
not use codes as identifiers, but instead use serial numbers. Set to TRUE if
these are not identified correctly.}
}
\value{
A tibble with the contents of the pxweb table.
}
\description{
Retrieve data from a pxweb database
}
\examples{
url_text <- pxw_create_api_url(paste0(
    "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__",
    "BE__BE0101__BE0101A/BefolkningNy/"
))


pxw_fetch(
    url_text = url_text, filters_list = list(Region = "1280", Tid = "2023",
                                             Civilstand = "e",
                                             Alder = "e"),
   kod_kolumn = "region"
)
}
