% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_first_period}
\alias{pxw_get_first_period}
\title{Get the first time in the table (year, month, etc.)}
\usage{
pxw_get_first_period(url_text)
}
\arguments{
\item{url_text}{URL to the tables web site or API.}
}
\value{
A text string with the earliest time found in the table
}
\description{
Get the first time in the table (year, month, etc.)
}
\examples{
url_text <- paste0(
  "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/",
  "START__AM__AM0210__AM0210A/ArbStatusM/"
)
pxw_get_first_period(url_text)
}
