% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_period_intervall}
\alias{pxw_get_period_intervall}
\title{Skapa ett periodintervall för de tidpunkter som finns i tabelan}
\usage{
pxw_get_period_intervall(url, from_per = "2020M06", to_per = "2024M01")
}
\arguments{
\item{url}{URL till tabellen webbsida eller API.}

\item{from_per}{Ange den tidigaste tidpunkten i intervallet}

\item{to_per}{Ange senaste tidpunkten i intervallet.}
}
\value{
En vektor med alla de tider som finns mellan from_per och to_per.
}
\description{
Skapa ett periodintervall för de tidpunkter som finns i tabelan
}
\examples{
url <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_get_period_intervall(url, from_per = NULL, to_per = NULL)
}
