% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_period_intervall}
\alias{pxw_get_period_intervall}
\title{Create a period range for the times in the table}
\usage{
pxw_get_period_intervall(url_text, from_per = NULL, to_per = NULL)
}
\arguments{
\item{url_text}{URL to the tables web site or API.}

\item{from_per}{Enter the earliest time in the range}

\item{to_per}{Enter the latest time in the range.}
}
\value{
A vector with all the times between from_per and to_per.
}
\description{
Create a period range for the times in the table
}
\examples{
url_text <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_get_period_intervall(url_text, from_per = "2024M01", to_per = "2024M03")
}
