% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_periods}
\alias{pxw_get_periods}
\title{Retrieve all available periods from the table}
\usage{
pxw_get_periods(url_text)
}
\arguments{
\item{url_text}{URL to the table's website or API.}
}
\value{
A vector containing all available times in the database, such as all
years for which data is available.
}
\description{
Retrieve all available periods from the table
}
\examples{
url_text <- paste0(
  "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/",
  "START__AM__AM0210__AM0210A/ArbStatusM/"
)
pxw_get_periods(url_text)
}
