% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_periods}
\alias{pxw_get_periods}
\title{Get all available periods}
\usage{
pxw_get_periods(url_text)
}
\arguments{
\item{url_text}{URL to the tables web site or API.}
}
\value{
A vector of all times available in the database, for example all
years for which there is data
}
\description{
Get all available periods
}
\examples{
url_text <- paste0(
  "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/",
  "START__AM__AM0210__AM0210A/ArbStatusM/"
)
pxw_get_periods(url_text)
}
