% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_get_region_codes}
\alias{pxw_get_region_codes}
\title{Om tabellen har en geografisk variabel hämtas koder för denna, till exempel
kommunkoder}
\usage{
pxw_get_region_codes(url, niva_nchar = NULL, lan_kod = NULL)
}
\arguments{
\item{url}{URL till tabellen, antingen på webbsajt eller api-URL}

\item{niva_nchar}{Om tabellen innehåller blandade geografiska nivåer så kan
man ange att enbart koder med en viss längd ska hämtas. niva_nchar = 2 hämtar
till exempel till exempel enbart länskoder.}

\item{lan_kod}{Vektor med länskoder. Om lan_kod = c("10", "12") så hämtas koder enbart för Blekinge och Skåne.}
}
\value{
En vektor med kommunkoder eller dylikt
}
\description{
Om tabellen har en geografisk variabel hämtas koder för denna, till exempel
kommunkoder
}
\examples{
url <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_get_region_codes(url, niva_nchar = 4, lan_kod = "12")
}
