% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_skapa_api_url}
\alias{pxw_skapa_api_url}
\title{Convert a link to a database tables home page to an api url}
\usage{
pxw_skapa_api_url(url_text)
}
\arguments{
\item{url_text}{An URL to a tables as a character string.}
}
\value{
A URL to a API
}
\description{
Convert a link to a database tables home page to an api url
}
\examples{
url_text <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_skapa_api_url(url_text)
}
