% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_variables_list}
\alias{pxw_variables_list}
\title{Retrieve a tibble with all parameters and variables from the table}
\usage{
pxw_variables_list(url_text)
}
\arguments{
\item{url_text}{URL to the table's website or API.}
}
\value{
A tibble containing all parameters and variables available for querying.
}
\description{
Retrieve a tibble with all parameters and variables from the table
}
\examples{
url_text <- paste0(
  "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/",
  "START__AM__AM0210__AM0210A/ArbStatusM/"
)
pxw_variables_list(url_text)
}
