% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_variables_list_mod.R
\name{pxw_variables_list_mod}
\alias{pxw_variables_list_mod}
\title{Adjust variable lists from misconfigured pxweb servers to follow SCB standards}
\usage{
pxw_variables_list_mod(url_str, lopnr_in_values = FALSE)
}
\arguments{
\item{url_str}{A string representing the URL to a pxweb API.}

\item{lopnr_in_values}{A logical value; set to TRUE if the server does not
have categories in the value parameter.}
}
\value{
A dataframe containing all parameters and variables in use.
}
\description{
Adjust variable lists from misconfigured pxweb servers to follow SCB standards
}
\examples{
link <- paste0("https://statistik.tillvaxtanalys.se/PxWeb/pxweb/sv/",
"Tillv\%C3\%A4xtanalys\%20statistikdatabas/Tillv\%C3\%A4xtanalys",
"\%20statistikdatabas__Konkurser\%20och\%20offentliga\%20ackord/",
"konk_ar_lan_bransch_2009.px/")
pxw_variables_list_mod(link)
}
