% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxw_fetch.R
\name{pxw_visa_parameterlista}
\alias{pxw_visa_parameterlista}
\title{Get a tibble with all parameters in the table}
\usage{
pxw_visa_parameterlista(url_text)
}
\arguments{
\item{url_text}{URL to the tables web site or API.}
}
\value{
A tibble with all parameters
}
\description{
Get a tibble with all parameters in the table
}
\examples{
url_text <- "https://www.statistikdatabasen.scb.se/pxweb/sv/ssd/START__AM__AM0210__AM0210A/ArbStatusM/"
pxw_visa_parameterlista(url_text)
}
